+++ src/Protocol/Planning/EntityQueryPlan.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace LaravelUi5\OData\Protocol\Planning;
 
 use LaravelUi5\OData\Edm\Contracts\Container\EntitySetInterface;
+use LaravelUi5\OData\Http\CustomQueryOptions;
 
 final readonly class EntityQueryPlan extends QueryPlan
 {
@@ -14,5 +15,6 @@ final readonly class EntityQueryPlan extends QueryPlan
         public SelectList         $select,
         public ExpandList         $expand,
         public ?NavigationAnchor  $anchor = null,
+        public CustomQueryOptions $customQueryOptions = new CustomQueryOptions(),
     ) {}
 }