+++ tests-fixtures/PassengerColours.php
@@ -6,6 +6,7 @@ namespace LaravelUi5\OData\Fixtures;
 
 use Illuminate\Database\Query\Builder;
 use Illuminate\Support\Facades\DB;
+use LaravelUi5\OData\Http\CustomQueryOptions;
 use LaravelUi5\OData\Edm\EdmPrimitiveType;
 use LaravelUi5\OData\Fixtures\Models\Enums\Colour;
 use LaravelUi5\OData\Service\AbstractEntitySet;
@@ -38,7 +39,7 @@ final readonly class PassengerColours extends AbstractEntitySet
         ];
     }
 
-    public function query(): Builder
+    public function query(CustomQueryOptions $options): Builder
     {
         return DB::table('passengers')->select('id', 'name', 'colour');
     }